rkyv::collections::btree_map

Struct ArchivedBTreeMap

pub struct ArchivedBTreeMap<K, V, const E: usize = 5> { /* private fields */ }

An archived BTreeMap.

Implementations

impl<K, V, const E: usize> ArchivedBTreeMap<K, V, E>

pub fn iter(&self) -> Iter<'_, K, V, E>

Gets an iterator over the entries of the map, sorted by key.

pub fn iter_seal(this: Seal<'_, Self>) -> IterSeal<'_, K, V, E>

Gets a mutable iterator over the entires of the map, sorted by key.

pub fn keys(&self) -> Keys<'_, K, V, E>

Gets an iterator over the sorted keys of the map.

pub fn values(&self) -> Values<'_, K, V, E>

Gets an iterator over the values of the map.

pub fn values_seal(this: Seal<'_, Self>) -> ValuesSeal<'_, K, V, E>

Gets a mutable iterator over the values of the map.

pub fn range<Q, R>(&self, range: R) -> Range<'_, K, V, E>
where Q: Ord + ?Sized, R: RangeBounds<Q>, K: Borrow<Q> + Ord,

Gets an iterator over a sub-range of entries, sorted by key.

pub fn range_with<Q, R, C>(&self, range: R, cmp: C) -> Range<'_, K, V, E>
where Q: Ord + ?Sized, R: RangeBounds<Q>, C: Fn(&Q, &K) -> Ordering, K: Ord,

Gets an iterator over a sub-range of entries, sorted by key.

This method uses the supplied comparison function to compare the range to elements.

pub fn range_seal<Q, R>( this: Seal<'_, Self>, range: R, ) -> RangeSeal<'_, K, V, E>
where Q: Ord + ?Sized, R: RangeBounds<Q>, K: Borrow<Q> + Ord,

Gets a mutable iterator over a sub-range of entries, sorted by key.

pub fn range_seal_with<Q, R, C>( this: Seal<'_, Self>, range: R, cmp: C, ) -> RangeSeal<'_, K, V, E>
where Q: Ord + ?Sized, R: RangeBounds<Q>, C: Fn(&Q, &K) -> Ordering, K: Ord,

Gets a mutable iterator over a sub-range of entries, sorted by key.

This method uses the supplied comparison function to compare the range to elements.

impl<K, V, const E: usize> ArchivedBTreeMap<K, V, E>

pub fn contains_key<Q>(&self, key: &Q) -> bool
where Q: Ord + ?Sized, K: Borrow<Q> + Ord,

Returns whether the B-tree map contains the given key.

pub fn get<Q>(&self, key: &Q) -> Option<&V>
where Q: Ord + ?Sized, K: Borrow<Q> + Ord,

Returns the value associated with the given key, or None if the key is not present in the B-tree map.

pub fn get_seal<'a, Q>(this: Seal<'a, Self>, key: &Q) -> Option<Seal<'a, V>>
where Q: Ord + ?Sized, K: Borrow<Q> + Ord,

Returns the mutable value associated with the given key, or None if the key is not present in the B-tree map.

pub fn is_empty(&self) -> bool

Returns true if the B-tree map contains no entries.

pub fn len(&self) -> usize

Returns the number of entries in the B-tree map.

pub fn get_key_value<Q>(&self, key: &Q) -> Option<(&K, &V)>
where Q: Ord + ?Sized, K: Borrow<Q> + Ord,

Gets the key-value pair associated with the given key, or None if the key is not present in the B-tree map.

pub fn get_key_value_with<Q, C>(&self, key: &Q, cmp: C) -> Option<(&K, &V)>
where Q: Ord + ?Sized, C: Fn(&Q, &K) -> Ordering, K: Ord,

Gets the key-value pair associated with the given key, or None if the key is not present in the B-tree map.

This method uses the supplied comparison function to compare the key to elements.

pub fn get_key_value_seal<'a, Q>( this: Seal<'a, Self>, key: &Q, ) -> Option<(&'a K, Seal<'a, V>)>
where Q: Ord + ?Sized, K: Borrow<Q> + Ord,

Gets the mutable key-value pair associated with the given key, or None if the key is not present in the B-tree map.

pub fn get_key_value_seal_with<'a, Q, C>( this: Seal<'a, Self>, key: &Q, cmp: C, ) -> Option<(&'a K, Seal<'a, V>)>
where Q: Ord + ?Sized, C: Fn(&Q, &K) -> Ordering, K: Ord,

Gets the mutable key-value pair associated with the given key, or None if the key is not present in the B-tree map.

This method uses the supplied comparison function to compare the key to elements.

pub fn resolve_from_len( len: usize, resolver: BTreeMapResolver, out: Place<Self>, )

Resolves an ArchivedBTreeMap from the given length, resolver, and output place.

pub fn serialize_from_ordered_iter<I, BKU, BVU, KU, VU, S>( iter: I, serializer: &mut S, ) -> Result<BTreeMapResolver, S::Error>
where I: ExactSizeIterator<Item = (BKU, BVU)>, BKU: Borrow<KU>, BVU: Borrow<VU>, KU: Serialize<S, Archived = K>, VU: Serialize<S, Archived = V>, S: Fallible + Allocator + Writer + ?Sized, S::Error: Source,

Serializes an ArchivedBTreeMap from the given iterator and serializer.

pub fn visit<T>(&self, f: impl FnMut(&K, &V) -> ControlFlow<T>) -> Option<T>

Visits every key-value pair in the B-tree with a function.

If f returns ControlFlow::Break, visit will return Some with the broken value. If f returns Continue for every pair in the tree, visit will return None.

pub fn visit_seal<T>( this: Seal<'_, Self>, f: impl FnMut(&K, Seal<'_, V>) -> ControlFlow<T>, ) -> Option<T>

Visits every mutable key-value pair in the B-tree with a function.

If f returns ControlFlow::Break, visit will return Some with the broken value. If f returns Continue for every pair in the tree, visit will return None.