Module rkyv::collections::hash_set

Archived hash set implementation.

During archiving, hashsets are built into minimal perfect hashsets using compress, hash and displace.

Structs

ArchivedHashSet

An archived HashSet. This is a wrapper around a hash map with the same key and a value of ().

HashSetResolver

The resolver for archived hash sets.