Struct rkyv::collections::hash_set::ArchivedHashSet

#[repr(transparent)]
pub struct ArchivedHashSet<K>(_);

An archived HashSet. This is a wrapper around a hash map with the same key and a value of ().

Implementations

impl<K> ArchivedHashSet<K>

pub const fn len(&self) -> usize

Gets the number of items in the hash set.

pub fn get<Q: ?Sized>(&self, k: &Q) -> Option<&K> where
K: Borrow<Q>,
Q: Hash + Eq,

Gets the key corresponding to the given key in the hash set.

pub fn contains<Q: ?Sized>(&self, k: &Q) -> bool where
K: Borrow<Q>,
Q: Hash + Eq,

Returns whether the given key is in the hash set.

pub fn hasher(&self) -> SeaHasher

Gets the hasher for the underlying hash map.

pub const fn is_empty(&self) -> bool

Returns whether there are no items in the hash set.

pub fn iter(&self) -> Keys<'_, K, ()>

Gets an iterator over the keys of the underlying hash map.

pub unsafe fn resolve_from_len(
len: usize,
pos: usize,
resolver: HashSetResolver,
out: *mut Self
)

Resolves an archived hash set from the given length and parameters.

Safety

len must be the number of elements that were serialized

pos must be the position of out within the archive

resolver must be the result of serializing a hash map

pub unsafe fn serialize_from_iter<'a, KU, S, I>(
iter: I,
serializer: &mut S
) -> Result<HashSetResolver, S::Error> where
KU: 'a + Serialize<S, Archived = K> + Hash + Eq,
S: Serializer + ScratchSpace + ?Sized,
I: ExactSizeIterator<Item = &'a KU>,

Serializes an iterator of keys as a hash set.

Safety

The keys returned by the iterator must be unique.