rkyv::collections::swiss_table::index_map

Struct ArchivedIndexMap

pub struct ArchivedIndexMap<K, V, H = FxHasher64> { /* private fields */ }

An archived IndexMap.

Implementations

impl<K, V, H> ArchivedIndexMap<K, V, H>

pub const fn is_empty(&self) -> bool

Returns true if the map contains no elements.

pub fn iter(&self) -> Iter<'_, K, V>

Returns an iterator over the key-value pairs of the map in order

pub fn keys(&self) -> Keys<'_, K, V>

Returns an iterator over the keys of the map in order

pub const fn len(&self) -> usize

Gets the number of items in the index map.

pub fn values(&self) -> Values<'_, K, V>

Returns an iterator over the values of the map in order.

impl<K, V, H: Hasher + Default> ArchivedIndexMap<K, V, H>

pub fn get_full_with<Q, C>(&self, key: &Q, cmp: C) -> Option<(usize, &K, &V)>
where Q: Hash + Eq + ?Sized, C: Fn(&Q, &K) -> bool,

Gets the index, key, and value corresponding to the supplied key using the given comparison function.

pub fn get_full<Q>(&self, key: &Q) -> Option<(usize, &K, &V)>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Gets the index, key, and value corresponding to the supplied key.

pub fn get_key_value_with<Q, C>(&self, key: &Q, cmp: C) -> Option<(&K, &V)>
where Q: Hash + Eq + ?Sized, C: Fn(&Q, &K) -> bool,

Returns the key-value pair corresponding to the supplied key using the given comparison function.

pub fn get_key_value<Q>(&self, key: &Q) -> Option<(&K, &V)>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Returns the key-value pair corresponding to the supplied key.

pub fn get_with<Q, C>(&self, key: &Q, cmp: C) -> Option<&V>
where Q: Hash + Eq + ?Sized, C: Fn(&Q, &K) -> bool,

Returns a reference to the value corresponding to the supplied key using the given comparison function.

pub fn get<Q>(&self, key: &Q) -> Option<&V>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Returns a reference to the value corresponding to the supplied key.

pub fn get_full_seal_with<'a, Q, C>( this: Seal<'a, Self>, key: &Q, cmp: C, ) -> Option<(usize, &'a K, Seal<'a, V>)>
where Q: Hash + Eq + ?Sized, C: Fn(&Q, &K) -> bool,

Gets the mutable index, key, and value corresponding to the supplied key using the given comparison function.

pub fn get_full_seal<'a, Q>( this: Seal<'a, Self>, key: &Q, ) -> Option<(usize, &'a K, Seal<'a, V>)>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Gets the mutable index, key, and value corresponding to the supplied key.

pub fn get_key_value_seal_with<'a, Q, C>( this: Seal<'a, Self>, key: &Q, cmp: C, ) -> Option<(&'a K, Seal<'a, V>)>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized, C: Fn(&Q, &K) -> bool,

Returns the mutable key-value pair corresponding to the supplied key using the given comparison function.

pub fn get_key_value_seal<'a, Q>( this: Seal<'a, Self>, key: &Q, ) -> Option<(&'a K, Seal<'a, V>)>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Returns the mutable key-value pair corresponding to the supplied key.

pub fn get_seal_with<'a, Q, C>( this: Seal<'a, Self>, key: &Q, cmp: C, ) -> Option<Seal<'a, V>>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized, C: Fn(&Q, &K) -> bool,

Returns a mutable reference to the value corresponding to the supplied key using the given comparison function.

pub fn get_seal<'a, Q>(this: Seal<'a, Self>, key: &Q) -> Option<Seal<'a, V>>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Returns a mutable reference to the value corresponding to the supplied key.

pub fn contains_key<Q>(&self, key: &Q) -> bool
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Returns whether a key is present in the hash map.

pub fn get_index(&self, index: usize) -> Option<(&K, &V)>

Gets a key-value pair by index.

pub fn get_index_of_with<Q, C>(&self, key: &Q, cmp: C) -> Option<usize>
where Q: Hash + Eq + ?Sized, C: Fn(&Q, &K) -> bool,

Gets the index of a key if it exists in the map using the given comparison function.

pub fn get_index_of<Q>(&self, key: &Q) -> Option<usize>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Gets the index of a key if it exists in the map.

pub fn resolve_from_len( len: usize, load_factor: (usize, usize), resolver: IndexMapResolver, out: Place<Self>, )

Resolves an archived index map from a given length and parameters.

pub fn serialize_from_iter<I, BKU, BVU, KU, VU, S>( iter: I, load_factor: (usize, usize), serializer: &mut S, ) -> Result<IndexMapResolver, S::Error>
where I: Clone + ExactSizeIterator<Item = (BKU, BVU)>, BKU: Borrow<KU>, BVU: Borrow<VU>, KU: Serialize<S, Archived = K> + Hash + Eq, VU: Serialize<S, Archived = V>, S: Fallible + Writer + Allocator + ?Sized, S::Error: Source,

Serializes an iterator of key-value pairs as an index map.