rkyv::collections::swiss_table::table

Struct ArchivedHashTable

pub struct ArchivedHashTable<T> { /* private fields */ }

A low-level archived SwissTable hash table with explicit hashing.

Implementations

impl<T> ArchivedHashTable<T>

pub fn get_with<C>(&self, hash: u64, cmp: C) -> Option<&T>
where C: Fn(&T) -> bool,

Returns the key-value pair corresponding to the supplied key.

pub fn get_seal_with<C>( this: Seal<'_, Self>, hash: u64, cmp: C, ) -> Option<Seal<'_, T>>
where C: Fn(&T) -> bool,

Returns the mutable key-value pair corresponding to the supplied key.

pub const fn is_empty(&self) -> bool

Returns whether the hash table is empty.

pub const fn len(&self) -> usize

Returns the number of elements in the hash table.

pub fn capacity(&self) -> usize

Returns the total capacity of the hash table.

pub fn raw_iter(&self) -> RawIter<T>

Returns an iterator over the entry pointers in the hash table.

pub fn raw_iter_seal(this: Seal<'_, Self>) -> RawIter<T>

Returns a sealed iterator over the entry pointers in the hash table.

pub fn serialize_from_iter<I, U, H, S>( items: I, hashes: H, load_factor: (usize, usize), serializer: &mut S, ) -> Result<HashTableResolver, S::Error>
where I: Clone + ExactSizeIterator, I::Item: Borrow<U>, U: Serialize<S, Archived = T>, H: ExactSizeIterator<Item = u64>, S: Fallible + Writer + Allocator + ?Sized, S::Error: Source,

Serializes an iterator of items as a hash table.

pub fn resolve_from_len( len: usize, load_factor: (usize, usize), resolver: HashTableResolver, out: Place<Self>, )

Resolves an archived hash table from a given length and parameters.