Enum rkyv::core_impl::ArchivedOption

#[repr(u8)]
pub enum ArchivedOption<T> {
    None,
    Some(T),
}

An archived Option.

It functions identically to Option but has a different internal representation to allow for archiving.

Variants

None

No value

Some(T)

Tuple Fields

0: T

Some value T

Implementations

impl<T> ArchivedOption<T>

pub fn is_none(&self) -> bool

Returns true if the option is a None value.

pub fn is_some(&self) -> bool

Returns true if the option is a Some value.

pub fn as_ref(&self) -> Option<&T>

Converts to an Option<&T>.

pub fn as_mut(&mut self) -> Option<&mut T>

Converts to an Option<&mut T>.

pub fn get_or_insert(&mut self, v: T) -> &mut T

Inserts v into the option if it is None, then returns a mutable reference to the contained value.

pub fn get_or_insert_with<F: FnOnce() -> T>(&mut self, f: F) -> &mut T

Inserts a value computed from f into the option if it is None, then returns a mutable reference to the contained value.