Macro rkyv::to_archived

macro_rules! to_archived {
    ($expr : expr) => { ... };
}

Returns the archived value of the given archived primitive.

This macro is not needed for most use cases. Its primary purpose is to simultaneously:

Convert values from (potentially) different primitives to their archived counterparts

Allow transformation in const contexts

Prevent linter warnings from unused into() calls

Users should feel free to use the more ergonomic into() where appropriate.