Struct rkyv::niche::option_nonzero::ArchivedOptionNonZeroU16

pub struct ArchivedOptionNonZeroU16 { /* private fields */ }

A niched archived Option<NonZeroU16>

Implementations

impl ArchivedOptionNonZeroU16

pub fn is_none(&self) -> bool

Returns true if the option is a None value.

pub fn is_some(&self) -> bool

Returns true if the option is a Some value.

pub fn as_ref(&self) -> Option<&Archived<NonZeroU16>>

Converts to an Option<&Archived<NonZeroU16>>

pub fn as_mut(&mut self) -> Option<&mut Archived<NonZeroU16>>

Converts to an Option<&mut Archived<NonZeroU16>>

pub fn as_pin_ref(self: Pin<&Self>) -> Option<Pin<&Archived<NonZeroU16>>>

Converts from Pin<&ArchivedOptionNonZeroU16> to Option<Pin<&Archived<NonZeroU16>>>.

pub fn as_pin_mut( self: Pin<&mut Self>, ) -> Option<Pin<&mut Archived<NonZeroU16>>>

Converts from Pin<&mut ArchivedOptionNonZeroU16> to Option<Pin<&mut Archived<NonZeroU16>>>.

pub fn iter(&self) -> Iter<'_, Archived<NonZeroU16>>

Returns an iterator over the possibly contained value.

pub fn iter_mut(&mut self) -> IterMut<'_, Archived<NonZeroU16>>

Returns a mutable iterator over the possibly contained value.

pub fn get_or_insert(&mut self, v: NonZeroU16) -> &mut Archived<NonZeroU16>

Inserts v into the option if it is None, then returns a mutable reference to the contained value.

pub fn get_or_insert_with<F: FnOnce() -> NonZeroU16>( &mut self, f: F, ) -> &mut Archived<NonZeroU16>

Inserts a value computed from f into the option if it is None, then returns a mutable reference to the contained value.

pub unsafe fn resolve_from_option(field: Option<NonZeroU16>, out: *mut Self)

Resolves an ArchivedOptionNonZero from an Option<NonZero>.

Safety

pos must be the position of out within the archive