Enum rkyv::option::ArchivedOption

#[repr(u8)]
pub enum ArchivedOption<T> {
    None,
    Some(T),
}

An archived Option.

It functions identically to Option but has a different internal representation to allow for archiving.

Variants

None

No value

Some(T)

Tuple Fields

0: T

Some value T

Implementations

impl<T> ArchivedOption<T>

pub fn is_none(&self) -> bool

Returns true if the option is a None value.

pub fn is_some(&self) -> bool

Returns true if the option is a Some value.

pub const fn as_ref(&self) -> Option<&T>

Converts to an Option<&T>.

pub fn as_mut(&mut self) -> Option<&mut T>

Converts to an Option<&mut T>.

pub fn as_pin_ref(self: Pin<&Self>) -> Option<Pin<&T>>

Converts from Pin<&ArchivedOption<T>> to Option<Pin<&T>>.

pub fn as_pin_mut(self: Pin<&mut Self>) -> Option<Pin<&mut T>>

Converts from Pin<&mut ArchivedOption<T>> to Option<Pin<&mut T>>.

pub const fn iter(&self) -> Iter<'_, T>

Returns an iterator over the possibly contained value.

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Returns a mutable iterator over the possibly contained value.

pub fn get_or_insert(&mut self, v: T) -> &mut T

Inserts v into the option if it is None, then returns a mutable reference to the contained value.

pub fn get_or_insert_with<F: FnOnce() -> T>(&mut self, f: F) -> &mut T

Inserts a value computed from f into the option if it is None, then returns a mutable reference to the contained value.

impl<T: Deref> ArchivedOption<T>

pub fn as_deref(&self) -> Option<&<T as Deref>::Target>

Converts from &ArchivedOption<T> to Option<&T::Target>.

Leaves the original ArchivedOption in-place, creating a new one with a reference to the original one, additionally coercing the contents via Deref.

impl<T: DerefMut> ArchivedOption<T>

pub fn as_deref_mut(&mut self) -> Option<&mut <T as Deref>::Target>

Converts from &mut ArchivedOption<T> to Option<&mut T::Target>.

Leaves the original ArchivedOption in-place, creating a new Option with a mutable reference to the inner type’s Deref::Target type.