Struct rkyv::rc::ArchivedRc

pub struct ArchivedRc<T: ArchivePointee + ?Sized, F>(/* private fields */);

An archived Rc.

This is a thin wrapper around a RelPtr to the archived type paired with a “flavor” type. Because there may be many varieties of shared pointers and they may not be used together, the flavor helps check that memory is not being shared incorrectly during validation.

Implementations

impl<T: ArchivePointee + ?Sized, F> ArchivedRc<T, F>

pub fn get(&self) -> &T

Gets the value of the ArchivedRc.

pub unsafe fn get_pin_mut_unchecked(self: Pin<&mut Self>) -> Pin<&mut T>

Gets the pinned mutable value of this ArchivedRc.

Safety

Any other ArchivedRc pointers to the same value must not be dereferenced for the duration of the returned borrow.

pub unsafe fn resolve_from_ref<U: ArchiveUnsized<Archived = T> + ?Sized>( value: &U, pos: usize, resolver: RcResolver<MetadataResolver<U>>, out: *mut Self, )

Resolves an archived Rc from a given reference.

Safety

pos must be the position of out within the archive

resolver must be the result of serializing value

pub fn serialize_from_ref<U: SerializeUnsized<S> + ?Sized, S: Serializer + SharedSerializeRegistry + ?Sized>( value: &U, serializer: &mut S, ) -> Result<RcResolver<MetadataResolver<U>>, S::Error>

Serializes an archived Rc from a given reference.