Module rkyv::rel_ptr

Relative pointer implementations and options.

Structs

RawRelPtr

An untyped pointer which resolves relative to its position in memory.

RelPtr

A pointer which resolves to relative to its position in memory.

Enums

OffsetError

An error where the distance between two positions cannot be represented by the offset type.

RelPtrError

Errors that can occur while creating raw relative pointers.

Traits

Offset

A offset that can be used with RawRelPtr.

Functions

signed_offset

Calculates the offset between two positions as an isize.

Type Aliases

RawRelPtrI8

A raw relative pointer that uses an archived i8 as the underlying offset.

RawRelPtrI16

A raw relative pointer that uses an archived i16 as the underlying offset.

RawRelPtrI32

A raw relative pointer that uses an archived i32 as the underlying offset.

RawRelPtrI64

A raw relative pointer that uses an archived i64 as the underlying offset.

RawRelPtrU8

A raw relative pointer that uses an archived u8 as the underlying offset.

RawRelPtrU16

A raw relative pointer that uses an archived u16 as the underlying offset.

RawRelPtrU32

A raw relative pointer that uses an archived u32 as the underlying offset.

RawRelPtrU64

A raw relative pointer that uses an archived u64 as the underlying offset.