Enum rkyv::result::ArchivedResult

#[repr(u8)]
pub enum ArchivedResult<T, E> {
    Ok(T),
    Err(E),
}

An archived Result that represents either success (Ok) or failure (Err).

Variants

Ok(T)

Contains the success value

Err(E)

Contains the error value

Implementations

impl<T, E> ArchivedResult<T, E>

pub const fn is_ok(&self) -> bool

Returns true if the result is Ok.

pub const fn is_err(&self) -> bool

Returns true if the result is Err.

pub fn as_ref(&self) -> Result<&T, &E>

Returns a Result containing the success and error values of this ArchivedResult.

pub fn as_mut(&mut self) -> Result<&mut T, &mut E>

Converts from &mut ArchivedResult<T, E> to Result<&mut T, &mut E>.

pub fn iter(&self) -> Iter<'_, T>

Returns an iterator over the possibly contained value.

The iterator yields one value if the result is ArchivedResult::Ok, otherwise none.

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Returns a mutable iterator over the possibly contained value.

The iterator yields one value if the result is ArchivedResult::Ok, otherwise none.

impl<T: Deref, E> ArchivedResult<T, E>

pub fn as_deref(&self) -> Result<&<T as Deref>::Target, &E>

Converts from &ArchivedResult<T, E> to Result<&<T as Deref>::Target, &E>.

Coerces the Ok variant of the original ArchivedResult via Deref and returns the new Result.

impl<T: DerefMut, E> ArchivedResult<T, E>

pub fn as_deref_mut(&mut self) -> Result<&mut <T as Deref>::Target, &mut E>

Converts from &mut ArchivedResult<T, E> to Result<&mut <T as Deref>::Target, &mut E>.

Coerces the Ok variant of the original ArchivedResult via DerefMut and returns the new Result.