Struct rkyv::ser::adapters::SharedSerializerAdapter

pub struct SharedSerializerAdapter<S> { /* fields omitted */ }

An adapter that adds shared serialization support to a serializer.

Implementations

impl<S> SharedSerializerAdapter<S>

pub fn new(inner: S) -> Self

Wraps the given serializer and adds shared memory support.

pub fn into_inner(self) -> S

Consumes the adapter and returns the underlying serializer.