rkyv::ser::allocator

Struct Arena

pub struct Arena { /* private fields */ }

An arena allocator for allocations.

Reusing the same arena for multiple serializations will reduce the number of global allocations, which can save a considerable amount of time.

Implementations

impl Arena

pub const DEFAULT_CAPACITY: usize = 1024

The default capacity for arenas.

pub fn new() -> Self

Creates a new Arena with the default capacity.

pub fn with_capacity(cap: usize) -> Self

Creates a new Arena with at least the requested capacity.

pub fn shrink(&mut self) -> usize

Cleans up allocated blocks which are no longer in use.

The arena is automatically shrunk by acquire.

pub fn capacity(&self) -> usize

Returns the available capacity of the arena.

pub fn acquire(&mut self) -> ArenaHandle<'_>

Acquires a handle to the arena.

The returned handle has exclusive allocation rights in the arena.

pub fn into_raw(self) -> NonNull<()>

Consumes the Arena, returning a raw pointer.

pub unsafe fn from_raw(raw: NonNull<()>) -> Self

Constructs an arena from a raw pointer.

Safety

raw must have been returned from into_raw. from_raw takes ownership over the pointer, and so from_raw must not be called on the same pointer more than once.