rkyv::ser::serializers

Struct AlignedSerializer

pub struct AlignedSerializer<A> { /* private fields */ }

A serializer made specifically to work with AlignedVec.

This serializer makes it easier for the compiler to perform emplacement optimizations and may give better performance than a basic WriteSerializer.

Implementations

impl<A: Borrow<AlignedVec>> AlignedSerializer<A>

pub fn new(inner: A) -> Self

Creates a new AlignedSerializer by wrapping a Borrow<AlignedVec>.

pub fn into_inner(self) -> A

Consumes the serializer and returns the underlying type.