Struct rkyv::ser::serializers::BufferSerializer

pub struct BufferSerializer<T> { /* fields omitted */ }

Wraps a byte buffer and equips it with Serializer.

Common uses include archiving in #![no_std] environments and archiving small objects without allocating.

Examples

use rkyv::{
    archived_value,
    ser::{Serializer, serializers::BufferSerializer},
    Aligned,
    AlignedVec,
    Archive,
    Archived,
    Serialize,
};

#[derive(Archive, Serialize)]
enum Event {
    Spawn,
    Speak(String),
    Die,
}

let mut serializer = BufferSerializer::new(Aligned([0u8; 256]));
let pos = serializer.serialize_value(&Event::Speak("Help me!".to_string()))
    .expect("failed to archive event");
let buf = serializer.into_inner();
let archived = unsafe { archived_value::<Event>(buf.as_ref(), pos) };
if let Archived::<Event>::Speak(message) = archived {
    assert_eq!(message.as_str(), "Help me!");
} else {
    panic!("archived event was of the wrong type");
}

Implementations

impl<T> BufferSerializer<T>

pub fn new(inner: T) -> Self

Creates a new archive buffer from a byte buffer.

pub fn with_pos(inner: T, pos: usize) -> Self

Creates a new archive buffer from a byte buffer. The buffer will start writing at the given position, but the buffer must contain all bytes (otherwise the alignments of types may not be correct).

pub fn into_inner(self) -> T

Consumes the serializer and returns the underlying type.