Struct rkyv::ser::serializers::ScratchTracker

pub struct ScratchTracker<T> { /* private fields */ }

A passthrough scratch space allocator that tracks scratch space usage.

Implementations

impl<T> ScratchTracker<T>

pub fn new(inner: T) -> Self

Creates a new scratch tracker from the given inner scratch space.

pub fn max_bytes_allocated(&self) -> usize

Returns the maximum number of bytes that were concurrently allocated during serialization.

pub fn max_allocations(&self) -> usize

Returns the maximum number of concurrent allocations during serialization.

pub fn max_alignment(&self) -> usize

Returns the maximum alignment of scratch space requested during serialization.

pub fn min_buffer_size(&self) -> usize

Returns the minimum buffer size required to serialize the same data.

This calculation takes into account packing efficiency for slab allocated scratch space. It is not exact, and has an error bound of max_allocations * (max_alignment - 1) bytes. This should be suitably small for most use cases.

pub fn min_buffer_size_max_error(&self) -> usize

Returns the maximum error term for the minimum buffer size calculation.