Struct rkyv::ser::serializers::WriteSerializer

pub struct WriteSerializer<W: Write> { /* private fields */ }

Wraps a type that implements io::Write and equips it with Serializer.

Examples

use rkyv::ser::{serializers::WriteSerializer, Serializer};

let mut serializer = WriteSerializer::new(Vec::new());
assert_eq!(serializer.pos(), 0);
serializer.write(&[0u8, 1u8, 2u8, 3u8]);
assert_eq!(serializer.pos(), 4);
let buf = serializer.into_inner();
assert_eq!(buf.len(), 4);
assert_eq!(buf, vec![0u8, 1u8, 2u8, 3u8]);

Implementations

impl<W: Write> WriteSerializer<W>

pub fn new(inner: W) -> Self

Creates a new serializer from a writer.

pub fn with_pos(inner: W, pos: usize) -> Self

Creates a new serializer from a writer, and assumes that the underlying writer is currently at the given position.

pub fn into_inner(self) -> W

Consumes the serializer and returns the internal writer used to create it.