rkyv::ser::serializers
Type Alias AllocSerializer

pub type AllocSerializer<const N: usize> = CompositeSerializer<AlignedSerializer<AlignedVec>, FallbackScratch<HeapScratch<N>, AllocScratch>, SharedSerializeMap>;

A general-purpose serializer suitable for environments where allocations can be made.

AllocSerializer takes one argument: the amount of scratch space to allocate before spilling allocations over into heap memory. A large amount of scratch space may result in some of it not being used, but too little scratch space will result in many allocations and decreased performance. You should consider your use case carefully when determining how much scratch space to pre-allocate.

Aliased Type

struct AllocSerializer<const N: usize> { /* private fields */ }