Trait rkyv::ser::SeekSerializer

pub trait SeekSerializer: Serializer {
    fn seek(&mut self, pos: usize) -> Result<(), Self::Error>;

    fn serialize_front<T: Serialize<Self>>(
        &mut self, 
        value: &T
    ) -> Result<usize, Self::Error> { ... }

    fn serialize_unsized_front<T: SerializeUnsized<Self> + ?Sized>(
        &mut self, 
        value: &T
    ) -> Result<usize, Self::Error>
    where
        T::Metadata: Serialize<Self>,
        T::Archived: ArchivePointee<ArchivedMetadata = Archived<T::Metadata>>,
    { ... }
}

A serializer that can seek to an absolute position.

Required methods

fn seek(&mut self, pos: usize) -> Result<(), Self::Error>

Seeks the serializer to the given absolute position.

Provided methods

fn serialize_front<T: Serialize<Self>>(
&mut self,
value: &T
) -> Result<usize, Self::Error>

Archives the given value at the nearest available position. If the serializer is already aligned, it will archive it at the current position.

fn serialize_unsized_front<T: SerializeUnsized<Self> + ?Sized>(
&mut self,
value: &T
) -> Result<usize, Self::Error> where
T::Metadata: Serialize<Self>,
T::Archived: ArchivePointee<ArchivedMetadata = Archived<T::Metadata>>,

Archives a reference to the given value at the nearest available position. If the serializer is already aligned, it will archive it at the current position.