Module rkyv::std_impl::chd

Archive implementation for hashmaps.

During archiving, hashmaps are built into minimal perfect hashmaps using compress, hash and displace.

Structs

ArchivedHashMap
An archived HashMap.

ArchivedHashMapResolver
The resolver for archived hash maps.

ArchivedHashSet
An archived HashSet. This is a wrapper around a hash map with the same key and a value of ().

ArchivedHashSetResolver
The resolver for archived hash sets.

Iter
An iterator over the key-value pairs of a hash map.

IterPin
An iterator over the mutable key-value pairs of a hash map.

Keys
An iterator over the keys of a hash map.

Values
An iterator over the values of a hash map.

ValuesPin
An iterator over the mutable values of a hash map.