Struct rkyv::std_impl::chd::ArchivedHashMap

pub struct ArchivedHashMap<K, V> { /* fields omitted */ }

An archived HashMap.

Implementations

impl<K: Hash + Eq, V> ArchivedHashMap<K, V>

pub fn len(&self) -> usize

Gets the number of items in the hash map.

pub fn hasher(&self) -> SeaHasher

Gets the hasher for this hashmap. The hasher for all archived hashmaps is the same for reproducibility.

pub fn get_key_value<Q: ?Sized>(&self, k: &Q) -> Option<(&K, &V)> where
K: Borrow<Q>,
Q: Hash + Eq,

Finds the key-value entry for a key.

pub fn get_key_value_pin<Q: ?Sized>(
self: Pin<&mut Self>,
k: &Q
) -> Option<(&K, Pin<&mut V>)> where
K: Borrow<Q>,
Q: Hash + Eq,

Finds the mutable key-value entry for a key.

pub fn contains_key<Q: ?Sized>(&self, k: &Q) -> bool where
K: Borrow<Q>,
Q: Hash + Eq,

Returns whether a key is present in the hash map.

pub fn get<Q: ?Sized>(&self, k: &Q) -> Option<&V> where
K: Borrow<Q>,
Q: Hash + Eq,

Gets the value associated with the given key.

pub fn get_pin<Q: ?Sized>(self: Pin<&mut Self>, k: &Q) -> Option<Pin<&mut V>> where
K: Borrow<Q>,
Q: Hash + Eq,

Gets the mutable value associated with the given key.

pub fn is_empty(&self) -> bool

Returns whether there are no items in the hashmap.

pub fn iter(&self) -> Iter<'_, K, V>

Gets an iterator over the key-value entries in the hash map.

pub fn iter_pin(self: Pin<&mut Self>) -> IterPin<'_, K, V>

Gets an iterator over the mutable key-value entries in the hash map.

pub fn keys(&self) -> Keys<'_, K, V>

Gets an iterator over the keys in the hash map.

pub fn values(&self) -> Values<'_, K, V>

Gets an iterator over the values in the hash map.

pub fn values_pin(self: Pin<&mut Self>) -> ValuesPin<'_, K, V>

Gets an iterator over the mutable values in the hash map.

pub fn serialize_from_iter<'a, KU: 'a + Serialize<S, Archived = K> + Hash + Eq, VU: 'a + Serialize<S, Archived = V>, S: Serializer + ?Sized>(
iter: impl Iterator<Item = (&'a KU, &'a VU)>,
len: usize,
serializer: &mut S
) -> Result<ArchivedHashMapResolver, S::Error>

Serializes an archived hash map from an iterator