Struct rkyv::std_impl::ArchivedVec

#[repr(transparent)]
pub struct ArchivedVec<T>(_);

An archived Vec.

Uses a RelPtr to a T slice under the hood.

Implementations

impl<T> ArchivedVec<T>

pub fn as_slice(&self) -> &[T]

Gets the elements of the archived vec as a slice.

pub fn as_mut_slice(&mut self) -> &mut [T]

Gets the elements of the archived vec as a mutable slice.

pub fn index_pin<I>(
self: Pin<&mut Self>,
index: I
) -> Pin<&mut <[T] as Index<I>>::Output> where
[T]: IndexMut<I>,

Gets the element at the given index ot this archived vec as a pinned mutable reference.