Struct rkyv::RelPtr

pub struct RelPtr<T: ArchivePointee + ?Sized> { /* fields omitted */ }

A pointer which resolves to relative to its position in memory.

See Archive for an example of creating one.

Implementations

impl<T: ArchivePointee + ?Sized> RelPtr<T>

pub fn new(raw_ptr: RawRelPtr, metadata: T::ArchivedMetadata) -> Self

Creates a new relative pointer from the given raw pointer and metadata.

Safety

The caller must guarantee that:

raw_ptr is a valid relative pointer in its final position

raw_ptr points to a valid value

metadata is valid metadata for the pointed value.

pub fn resolve_emplace<U: ArchiveUnsized<Archived = T> + ?Sized>(
from: usize,
to: usize,
value: &U,
metadata_resolver: U::MetadataResolver,
out: &mut MaybeUninit<Self>
)

Creates a relative pointer from one position to another.

Safety

The caller must guarantee that from is the position of the relative pointer and to is the position of some valid memory.

pub fn base(&self) -> *const u8

Gets the base pointer for the relative pointer.

pub fn offset(&self) -> isize

Gets the offset of the relative pointer.

pub fn metadata(&self) -> &T::ArchivedMetadata

Gets the metadata of the relative pointer.

pub fn as_ptr(&self) -> *const T

Calculates the memory address being pointed to by this relative pointer.

pub fn as_mut_ptr(&mut self) -> *mut T

Returns an unsafe mutable pointer to the memory address being pointed to by this relative pointer.