Trait rkyv::ArchivePointee

pub trait ArchivePointee: Pointee {
    type ArchivedMetadata;
    fn pointer_metadata(
        archived: &Self::ArchivedMetadata
    ) -> <Self as Pointee>::Metadata;
}

An archived type with associated metadata for its relative pointer.

This is mostly used in the context of smart pointers and unsized types, and is implemented for all sized types by default.

Associated Types

type ArchivedMetadata

The archived version of the pointer metadata for this type.

Required methods

fn pointer_metadata(
archived: &Self::ArchivedMetadata
) -> <Self as Pointee>::Metadata

Converts some archived metadata to the pointer metadata for itself.