rkyv

Trait Portable

pub unsafe trait Portable { }

A type with a stable, well-defined layout that is the same on all targets.

Safety

The implementing type must have a stable, well-defined layout that is the same on all targets. Structs and unions must be #[repr(transparent)] or #[repr(C)]. Enums must be #[repr(C)], #[repr(int)], or #[repr(C, int)].

The implementing type must not have interior mutability (i.e. no UnsafeCells).