Type Definition rkyv::Archived

pub type Archived<T> = <T as Archive>::Archived;

Alias for the archived version of some Archive type.

This can be useful for reducing the lengths of type definitions.