rkyv::util

Function archived_root

pub unsafe fn archived_root<T: Archive + ?Sized>(bytes: &[u8]) -> &T::Archived

Casts an archived value from the given byte slice by calculating the root position.

This is a wrapper for archived_value that calculates the correct position of the root using the length of the byte slice. If your byte slice is not guaranteed to end immediately after the root object, you may need to store the position of the root object returned from serialize_value.

Safety

The byte slice must represent an archived object

The root of the object must be stored at the end of the slice (this is the default behavior)