Type Alias rkyv::validation::owned::CheckOwnedPointerError

pub type CheckOwnedPointerError<T, C> = OwnedPointerError<<<T as ArchivePointee>::ArchivedMetadata as CheckBytes<C>>::Error, <T as CheckBytes<C>>::Error, <C as Fallible>::Error>;

The OwnedPointerError for an owned T being checked with a some context C.

Aliased Type

enum CheckOwnedPointerError<T, C> {
    PointerCheckBytesError(Infallible),
    ValueCheckBytesError(<T as CheckBytes<C>>::Error),
    ContextError(<C as Fallible>::Error),
}

Variants

PointerCheckBytesError(Infallible)

The pointer failed to validate due to invalid metadata.

ValueCheckBytesError(<T as CheckBytes<C>>::Error)

The value pointed to by the owned pointer was invalid.

ContextError(<C as Fallible>::Error)

An error occurred from the validation context.