Module rkyv::validation::validators

Validators that can check archived types.

Structs

ArchiveValidator

A validator that can verify archives with nonlocal memory.

DefaultValidator

The default validator.

PrefixRange

A prefix range from an ArchiveValidator.

SharedValidator

A validator that can verify shared memory.

SuffixRange

A suffix range from an ArchiveValidator.

Enums

ArchiveError

Errors that can occur when checking archive memory.

CheckDeserializeError

Errors that can occur while deserializing from bytes.

DefaultValidatorError

The default validator error.

SharedError

Errors that can occur when checking shared memory.

Functions

check_archived_root

Checks the given archive at the given position for an archived version of the given type.

check_archived_value

Checks the given archive at the given position for an archived version of the given type.

from_bytes

Checks and deserializes a value from the given bytes.

Type Aliases

FromBytesError

The error type for from_bytes.