Enum rkyv::validation::validators::ArchiveError

pub enum ArchiveError {
    Overflow {
        base: *const u8,
        offset: isize,
    },
    Underaligned {
        expected_align: usize,
        actual_align: usize,
    },
    OutOfBounds {
        base: *const u8,
        offset: isize,
        range: Range<*const u8>,
    },
    Overrun {
        ptr: *const u8,
        size: usize,
        range: Range<*const u8>,
    },
    Unaligned {
        ptr: *const u8,
        align: usize,
    },
    SubtreePointerOutOfBounds {
        ptr: *const u8,
        subtree_range: Range<*const u8>,
    },
    SubtreePointerOverrun {
        ptr: *const u8,
        size: usize,
        subtree_range: Range<*const u8>,
    },
    RangePoppedOutOfOrder {
        expected_depth: usize,
        actual_depth: usize,
    },
    UnpoppedSubtreeRanges {
        last_range: usize,
    },
    ExceededMaximumSubtreeDepth {
        max_subtree_depth: usize,
    },
    LayoutError {
        layout_error: LayoutError,
    },
}

Errors that can occur when checking archive memory.

Variants

Overflow

Computing the target of a relative pointer overflowed

Fields

base: *const u8

The base pointer

offset: isize

The offset

Underaligned

The archive is under-aligned for one of the types inside

Fields

expected_align: usize

The expected alignment of the archive

actual_align: usize

The actual alignment of the archive

OutOfBounds

A pointer pointed outside the bounds of the archive

Fields

base: *const u8

The base of the relative pointer

offset: isize

The offset of the relative pointer

range: Range<*const u8>

The pointer range of the archive

Overrun

There wasnโt enough space for the desired type at the pointed location

Fields

ptr: *const u8

The pointer to the type

size: usize

The desired size of the type

range: Range<*const u8>

The pointer range of the archive

Unaligned

The pointer wasnโt aligned properly for the desired type

Fields

ptr: *const u8

The pointer to the type

align: usize

The required alignment of the type

SubtreePointerOutOfBounds

The pointer wasnโt within the subtree range

Fields

ptr: *const u8

The pointer to the subtree

subtree_range: Range<*const u8>

The subtree range

SubtreePointerOverrun

There wasnโt enough space in the subtree range for the desired type at the pointed location

Fields

ptr: *const u8

The pointer to the subtree type,

size: usize

The desired size of the type

subtree_range: Range<*const u8>

The subtree range

RangePoppedOutOfOrder

A subtree range was popped out of order.

Subtree ranges must be popped in the reverse of the order they are pushed.

Fields

expected_depth: usize

The expected depth of the range

actual_depth: usize

The actual depth of the range

UnpoppedSubtreeRanges

A subtree range was not popped before validation concluded.

Fields

last_range: usize

The depth of the last subtree that was pushed

ExceededMaximumSubtreeDepth

The maximum subtree depth was reached or exceeded.

Fields

max_subtree_depth: usize

The maximum depth that subtrees may be validated down to

LayoutError

A layout error occurred

Fields

layout_error: LayoutError

A layout error