Struct rkyv::vec::ArchivedVec

#[repr(C)]
pub struct ArchivedVec<T> { /* private fields */ }

An archived Vec.

This uses a RelPtr to a [T] under the hood. Unlike ArchivedString, it does not have an inline representation.

Implementations

impl<T> ArchivedVec<T>

pub fn as_ptr(&self) -> *const T

Returns a pointer to the first element of the archived vec.

pub fn len(&self) -> usize

Returns the number of elements in the archived vec.

pub fn is_empty(&self) -> bool

Returns whether the archived vec is empty.

pub fn as_slice(&self) -> &[T]

Gets the elements of the archived vec as a slice.

pub fn pin_mut_slice(self: Pin<&mut Self>) -> Pin<&mut [T]>

Gets the elements of the archived vec as a pinned mutable slice.

pub fn index_pin<I>( self: Pin<&mut Self>, index: I, ) -> Pin<&mut <[T] as Index<I>>::Output>
where [T]: IndexMut<I>,

Gets the element at the given index ot this archived vec as a pinned mutable reference.

pub unsafe fn resolve_from_slice<U: Archive<Archived = T>>( slice: &[U], pos: usize, resolver: VecResolver, out: *mut Self, )

Resolves an archived Vec from a given slice.

Safety

pos must be the position of out within the archive

resolver must be the result of serializing value

pub unsafe fn resolve_from_len( len: usize, pos: usize, resolver: VecResolver, out: *mut Self, )

Resolves an archived Vec from a given length.

Safety

pos must be the position of out within the archive

resolver must bet he result of serializing value

pub fn serialize_from_slice<U: Serialize<S, Archived = T>, S: Serializer + ?Sized>( slice: &[U], serializer: &mut S, ) -> Result<VecResolver, S::Error>
where [U]: SerializeUnsized<S>,

Serializes an archived Vec from a given slice.

pub unsafe fn serialize_copy_from_slice<U, S>( slice: &[U], serializer: &mut S, ) -> Result<VecResolver, S::Error>
where U: Serialize<S, Archived = T>, S: Serializer + ?Sized,

Serializes an archived Vec from a given slice by directly copying bytes.

Safety

The type being serialized must be copy-safe. Copy-safe types must be trivially copyable (have the same archived and unarchived representations) and contain no padding bytes. In situations where copying uninitialized bytes the output is acceptable, this function may be used with types that contain padding bytes.

pub fn serialize_from_iter<U, B, I, S>( iter: I, serializer: &mut S, ) -> Result<VecResolver, S::Error>
where U: Serialize<S, Archived = T>, B: Borrow<U>, I: ExactSizeIterator<Item = B>, S: ScratchSpace + Serializer + ?Sized,

Serializes an archived Vec from a given iterator.

This method is unable to perform copy optimizations; prefer serialize_from_slice when possible.

impl<T> ArchivedVec<T>

pub unsafe fn check_bytes_with<'a, C, F>( value: *const Self, context: &mut C, check_elements: F, ) -> Result<&'a Self, CheckOwnedPointerError<[T], C>>
where T: CheckBytes<C>, C: ArchiveContext + ?Sized, F: FnOnce(*const [T], &mut C) -> Result<(), <[T] as CheckBytes<C>>::Error>,

Checks the bytes of the ArchivedVec with the given element checking function.

Safety

check_elements must ensure that the pointer given to it contains only valid data.