Struct rkyv::with::RefAsBox

pub struct RefAsBox;

A wrapper that serializes a reference as if it were boxed.

Unlike Inline, unsized references can be serialized with RefAsBox.

References serialized with RefAsBox cannot be deserialized because the struct cannot own the deserialized value.

Example

use rkyv::{Archive, with::RefAsBox};

#[derive(Archive)]
struct Example<'a> {
    #[with(RefAsBox)]
    a: &'a i32,
    #[with(RefAsBox)]
    b: &'a str,
}