Trait rkyv::with::DeserializeWith

pub trait DeserializeWith<F: ?Sized, T, D: Fallible + ?Sized> {
    fn deserialize_with(field: &F, deserializer: &mut D) -> Result<T, D::Error>;
}

A variant of Deserialize that works with With wrappers.

Required methods

fn deserialize_with(field: &F, deserializer: &mut D) -> Result<T, D::Error>

Deserializes the field type F using the given deserializer.