Module rkyv::collections::btree_map

Archive implementation for B-tree maps.

Structs

ArchivedBTreeMap

An archived BTreeMap.

BTreeMapResolver

The resolver for an ArchivedBTreeMap.

Iter

An iterator over the key-value pairs of an archived B-tree map.

Keys

An iterator over the keys of an archived B-tree map.

Values

An iterator over the values of an archived B-tree map.

Constants

MIN_ENTRIES_PER_INNER_NODE

The minimum number of entries to place in an inner node.

MIN_ENTRIES_PER_LEAF_NODE

The minimum number of entries to place in a leaf node.