Struct rkyv::collections::btree_map::ArchivedBTreeMap

pub struct ArchivedBTreeMap<K, V> { /* fields omitted */ }

An archived BTreeMap.

Implementations

impl<K, V> ArchivedBTreeMap<K, V>

pub fn contains_key<Q: Ord + ?Sized>(&self, key: &Q) -> bool where
K: Borrow<Q> + Ord,

Returns true if the map contains a value for the specified key.

The key may be any borrowed form of the map’s key type, but the ordering on the borrowed form must match the ordering on the key type.

pub fn get<Q: Ord + ?Sized>(&self, key: &Q) -> Option<&V> where
K: Borrow<Q> + Ord,

Returns a reference to the value corresponding to the key.

The key may be any borrowed form of the map’s key type, but the ordering on the borrowed form must match the ordering on the key type.

pub fn get_key_value<Q: Ord + ?Sized>(&self, k: &Q) -> Option<(&K, &V)> where
K: Borrow<Q> + Ord,

Returns the key-value pair corresponding to the supplied key.

The supplied key may be any borrowed form of the map’s key type, but the ordering on the borrowed form must match the ordering on the key type.

pub fn is_empty(&self) -> bool

Returns true if the map contains no elements.

pub fn iter(&self) -> Iter<'_, K, V>

Gets an iterator over the entries of the map, sorted by key.

pub fn keys(&self) -> Keys<'_, K, V>

Gets an iterator over the keys of the map, in sorted order.

pub fn len(&self) -> usize

Returns the number of items in the archived B-tree map.

pub fn values(&self) -> Values<'_, K, V>

Gets an iterator over the values of the map, in order by key.

pub unsafe fn resolve_from_len(
len: usize,
pos: usize,
resolver: BTreeMapResolver,
out: *mut Self
)

Resolves a B-tree map from its length.

Safety

len must be the number of elements that were serialized

pos must be the position of out within the archive

resolver must be the result of serializing a B-tree map