Struct rkyv::collections::btree_set::ArchivedBTreeSet

#[repr(transparent)]
pub struct ArchivedBTreeSet<K>(_);

An archived BTreeSet. This is a wrapper around a B-tree map with the same key and a value of ().

Implementations

impl<K> ArchivedBTreeSet<K>

pub fn contains_key<Q: Ord + ?Sized>(&self, key: &Q) -> bool where
K: Borrow<Q> + Ord,

Returns true if the set contains a value for the specified key.

The key may be any borrowed form of the set’s key type, but the ordering on the borrowed form must match the ordering on the key type.

pub fn get<Q: Ord + ?Sized>(&self, value: &Q) -> Option<&K> where
K: Borrow<Q> + Ord,

Returns a reference to the value int he set, if any, that is equal to the given value.

The value may be any borrowed form of the set’s value type, but the ordering on the borrowed form must match the ordering on the value type.

pub fn is_empty(&self) -> bool

Returns true if the set contains no elements.

pub fn iter(&self) -> Keys<'_, K, ()>

Gets an iterator over the keys of the set, in sorted order.

pub fn len(&self) -> usize

Returns the number of items in the archived B-tree set.

pub unsafe fn resolve_from_len(
len: usize,
pos: usize,
resolver: BTreeSetResolver,
out: *mut Self
)

Resolves a B-tree set from its length.

Safety

len must be the number of elements that were serialized

pos must be the position of out within the archive

resolver must be the result of serializing a B-tree set