Trait rkyv::de::Deserializer

pub trait Deserializer: Fallible {
    unsafe fn alloc(&mut self, layout: Layout) -> Result<*mut u8, Self::Error>;
}

A context that provides a memory allocator.

Most types that support DeserializeUnsized will require this kind of context.

Required methods

unsafe fn alloc(&mut self, layout: Layout) -> Result<*mut u8, Self::Error>

Allocates and returns memory with the given layout.

Safety

The caller must guarantee that the memory returned by this function is deallocated by the global allocator.