Struct rkyv::ffi::ArchivedCString

pub struct ArchivedCString(/* private fields */);

An archived CString.

Uses a RelPtr to a CStr under the hood.

Implementations

impl ArchivedCString

pub fn as_bytes(&self) -> &[u8]

Returns the contents of this CString as a slice of bytes.

The returned slice does not contain the trailing nul terminator, and it is guaranteed to not have any interior nul bytes. If you need the nul terminator, use as_bytes_with_nul instead.

pub fn as_bytes_with_nul(&self) -> &[u8]

Equivalent to as_bytes except that the returned slice includes the trailing nul terminator.

pub fn as_c_str(&self) -> &CStr

Extracts a CStr slice containing the entire string.

pub fn pin_mut_c_str(self: Pin<&mut Self>) -> Pin<&mut CStr>

Extracts a pinned mutable CStr slice containing the entire string.

pub unsafe fn resolve_from_c_str( c_str: &CStr, pos: usize, resolver: CStringResolver, out: *mut Self, )

Resolves an archived C string from the given C string and parameters.

Safety

pos must be the position of out within the archive

resolver must be the result of serializing a C string

pub fn serialize_from_c_str<S: Serializer + ?Sized>( c_str: &CStr, serializer: &mut S, ) -> Result<CStringResolver, S::Error>

Serializes a C string.