Struct rkyv::niche::option_box::ArchivedOptionBox

#[repr(transparent)]
pub struct ArchivedOptionBox<T: ArchivePointee + ?Sized> { /* fields omitted */ }

A niched archived Option<Box<T>>.

It uses less space by storing the None variant as a null pointer.

Implementations

impl<T: ArchivePointee + ?Sized> ArchivedOptionBox<T>

pub fn is_none(&self) -> bool

Returns true if the option box is a None value.

pub fn is_some(&self) -> bool

Returns true if the option box is a Some value.

pub fn as_ref(&self) -> Option<&ArchivedBox<T>>

Converts to an Option<&ArchivedBox<T>>.

pub fn as_mut(&mut self) -> Option<&mut ArchivedBox<T>>

Converts to an Option<&mut ArchivedBox<T>>.

pub fn as_pin_ref(self: Pin<&Self>) -> Option<Pin<&ArchivedBox<T>>>

Converts from Pin<&ArchivedOptionBox<T>> to Option<Pin<&ArchivedBox<T>>>.

pub fn as_pin_mut(self: Pin<&mut Self>) -> Option<Pin<&mut ArchivedBox<T>>>

Converts from Pin<&mut ArchivedOption<T>> to Option<Pin<&mut ArchivedBox<T>>>.

pub fn iter(&self) -> Iter<'_, ArchivedBox<T>>

Returns an iterator over the possibly contained value.

pub fn iter_mut(&mut self) -> IterMut<'_, ArchivedBox<T>>

Returns a mutable iterator over the possibly contained value.

pub fn as_deref(&self) -> Option<&T>

Converts from &ArchivedOptionBox<T> to Option<&T>.

Leaves the original ArchivedOptionBox in-place, creating a new one with a reference to the original one.

impl<T: ArchivePointee + ?Sized> ArchivedOptionBox<T> where
T::ArchivedMetadata: Default,

pub unsafe fn resolve_from_option<U: ArchiveUnsized<Archived = T> + ?Sized>(
field: Option<&U>,
pos: usize,
resolver: OptionBoxResolver<U::MetadataResolver>,
out: *mut Self
)

Resolves an ArchivedOptionBox<T::Archived> from an Option<&T>.

Safety

pos must be the position of out within the archive

resolver must be the result of serializing field

pub fn serialize_from_option<U, S>(
field: Option<&U>,
serializer: &mut S
) -> Result<OptionBoxResolver<U::MetadataResolver>, S::Error> where
U: SerializeUnsized<S, Archived = T> + ?Sized,
S: Serializer + ?Sized,

Serializes an ArchivedOptionBox<T::Archived> from an Option<&T>.