rkyv::niche

Module option_nonzero

Niched archived Option<NonZero> integers that use less space.

Structs

ArchivedOptionNonZeroI8

A niched archived Option<NonZeroI8>

ArchivedOptionNonZeroI16

A niched archived Option<NonZeroI16>

ArchivedOptionNonZeroI32

A niched archived Option<NonZeroI32>

ArchivedOptionNonZeroI64

A niched archived Option<NonZeroI64>

ArchivedOptionNonZeroI128

A niched archived Option<NonZeroI128>

ArchivedOptionNonZeroU8

A niched archived Option<NonZeroU8>

ArchivedOptionNonZeroU16

A niched archived Option<NonZeroU16>

ArchivedOptionNonZeroU32

A niched archived Option<NonZeroU32>

ArchivedOptionNonZeroU64

A niched archived Option<NonZeroU64>

ArchivedOptionNonZeroU128

A niched archived Option<NonZeroU128>

Type Aliases

Iter

An iterator over a reference to the Some variant of an ArchivedOptionNonZero integer.

IterMut

An iterator over a mutable reference to the Some variant of an ArchivedOptionNonZero integer.