Struct rkyv::rel_ptr::RawRelPtr

pub struct RawRelPtr<O> { /* private fields */ }

An untyped pointer which resolves relative to its position in memory.

This is the most fundamental building block in rkyv. It allows the construction and use of pointers that can be safely relocated as long as the source and target are moved together. This is what allows memory to be moved from disk into memory and accessed without decoding.

Regular pointers are absolute, meaning that the pointee can be moved without invalidating the pointer. However, the target cannot be moved or the pointer is invalidated.

Relative pointers are relative, meaning that the pointee can be moved with the target without invalidating the pointer. However, if either the pointee or the target move independently, the pointer will be invalidated.

Implementations

impl<O: Offset> RawRelPtr<O>

pub unsafe fn try_emplace( from: usize, to: usize, out: *mut Self ) -> Result<(), OffsetError>

Attempts to create a new RawRelPtr in-place between the given from and to positions.

Safety

out must be located at position from

to must be a position within the archive

pub unsafe fn emplace(from: usize, to: usize, out: *mut Self)

Creates a new RawRelPtr in-place between the given from and to positions.

Safety

out must be located at position from

to must be a position within the archive

The offset between from and to must fit in an isize and not exceed the offset storage

pub fn base(&self) -> *const u8

Gets the base pointer for the relative pointer.

pub fn base_mut(&mut self) -> *mut u8

Gets the mutable base pointer for the relative pointer.

pub fn offset(&self) -> isize

Gets the offset of the relative pointer from its base.

pub fn is_null(&self) -> bool

Gets whether the offset of the relative pointer is 0.

pub fn as_ptr(&self) -> *const ()

Calculates the memory address being pointed to by this relative pointer.

pub fn as_mut_ptr(&mut self) -> *mut ()

Returns an unsafe mutable pointer to the memory address being pointed to by this relative pointer.