Struct rkyv::rel_ptr::RelPtr

pub struct RelPtr<T: ArchivePointee + ?Sized, O> { /* private fields */ }

A pointer which resolves to relative to its position in memory.

This is a strongly-typed version of RawRelPtr.

See Archive for an example of creating one.

Implementations

impl<T: ArchivePointee + ?Sized, O: Offset> RelPtr<T, O>

pub unsafe fn manual_check_bytes<'a, C: Fallible + ?Sized>( value: *const Self, context: &mut C, ) -> Result<&'a Self, <T::ArchivedMetadata as CheckBytes<C>>::Error>
where O: CheckBytes<C>, T::ArchivedMetadata: CheckBytes<C>,

Checks the bytes of the given relative pointer.

This is done rather than implementing CheckBytes to force users to manually write their CheckBytes implementation since they need to also provide the ownership model of their memory.

Safety

The given pointer must be aligned and point to enough bytes to represent a RelPtr<T>.

impl<T, O: Offset> RelPtr<T, O>

pub unsafe fn try_emplace( from: usize, to: usize, out: *mut Self, ) -> Result<(), OffsetError>

Attempts to create a relative pointer from one position to another.

Safety

from must be the position of out within the archive

to must be the position of some valid T

pub unsafe fn emplace(from: usize, to: usize, out: *mut Self)

Creates a relative pointer from one position to another.

Panics

The offset between from and to does not fit in an isize

The offset between from and to exceeds the offset storage

Safety

from must be the position of out within the archive

to must be the position of some valid T

impl<T: ArchivePointee + ?Sized, O: Offset> RelPtr<T, O>
where T::ArchivedMetadata: Default,

pub unsafe fn try_emplace_null( pos: usize, out: *mut Self, ) -> Result<(), OffsetError>

Attempts to create a null relative pointer with default metadata.

Safety

pos must be the position of out within the archive.

pub unsafe fn emplace_null(pos: usize, out: *mut Self)

Creates a null relative pointer with default metadata.

Panics

An offset of 0 does not fit in an isize

An offset of 0 exceeds the offset storage

Safety

pos must be the position of out within the archive.

impl<T: ArchivePointee + ?Sized, O: Offset> RelPtr<T, O>

pub unsafe fn try_resolve_emplace<U: ArchiveUnsized<Archived = T> + ?Sized>( from: usize, to: usize, value: &U, metadata_resolver: U::MetadataResolver, out: *mut Self, ) -> Result<(), OffsetError>

Attempts to create a relative pointer from one position to another.

Safety

from must be the position of out within the archive

to must be the position of some valid T

value must be the value being serialized

metadata_resolver must be the result of serializing the metadata of value

pub unsafe fn resolve_emplace<U: ArchiveUnsized<Archived = T> + ?Sized>( from: usize, to: usize, value: &U, metadata_resolver: U::MetadataResolver, out: *mut Self, )

Creates a relative pointer from one position to another.

Panics

The offset between from and to does not fit in an isize

The offset between from and to exceeds the offset storage

Safety

from must be the position of out within the archive

to must be the position of some valid T

value must be the value being serialized

metadata_resolver must be the result of serializing the metadata of value

pub unsafe fn try_resolve_emplace_from_raw_parts( from: usize, to: usize, archived_metadata: <T as ArchivePointee>::ArchivedMetadata, out: *mut Self, ) -> Result<(), OffsetError>

Attempts to create a relative pointer from one position to another given raw pointer metadata.

This does the same thing as RelPtr::try_resolve_emplace but you must supply the <T as ArchivePointee>::ArchivedMetadata yourself directly rather than through an implementation of ArchiveUnsized on some value.

Safety

from must be the position of out within the archive

to must be the position of some valid T

value must be the value being serialized

archived_metadata must produce valid metadata for the pointee of the resulting RelPtr (the thing being pointed at) when <T as ArchivePointee>::pointer_metadata(archived_metadata) is called.

pub unsafe fn resolve_emplace_from_raw_parts( from: usize, to: usize, archived_metadata: <T as ArchivePointee>::ArchivedMetadata, out: *mut Self, )

Creates a relative pointer from one position to another given raw pointer metadata.

This does the same thing as RelPtr::resolve_emplace but you must supply the <T as ArchivePointee>::ArchivedMetadata yourself directly rather than through an implementation of ArchiveUnsized on some value.

Panics

The offset between from and to does not fit in an isize

The offset between from and to exceeds the offset storage

Safety

from must be the position of out within the archive

to must be the position of some valid T

value must be the value being serialized

archived_metadata must produce valid metadata for the pointee of the resulting RelPtr (the thing being pointed at) when <T as ArchivePointee>::pointer_metadata(archived_metadata) is called.

pub fn base(&self) -> *const u8

Gets the base pointer for the relative pointer.

pub fn base_mut(&mut self) -> *mut u8

Gets the mutable base pointer for the relative pointer.

pub fn offset(&self) -> isize

Gets the offset of the relative pointer from its base.

pub fn is_null(&self) -> bool

Gets whether the offset of the relative pointer is 0.

pub fn metadata(&self) -> &T::ArchivedMetadata

Gets the metadata of the relative pointer.

pub fn as_ptr(&self) -> *const T

Calculates the memory address being pointed to by this relative pointer.

pub fn as_mut_ptr(&mut self) -> *mut T

Returns an unsafe mutable pointer to the memory address being pointed to by this relative pointer.