Type Alias rkyv::ser::serializers::CoreSerializer

pub type CoreSerializer<const S: usize, const C: usize> = CompositeSerializer<BufferSerializer<AlignedBytes<S>>, BufferScratch<AlignedBytes<C>>, Infallible>;

A serializer suitable for environments where allocations cannot be made.

CoreSerializer takes two arguments: the amount of serialization memory to allocate and the amount of scratch space to allocate. If you run out of either while serializing, the serializer will return an error.

Aliased Type

struct CoreSerializer<const S: usize, const C: usize> { /* private fields */ }