rkyv::string::repr

Union ArchivedStringRepr

pub union ArchivedStringRepr {
    /* private fields */
}

An archived string representation that can inline short strings.

Implementations

impl ArchivedStringRepr

pub fn is_inline(&self) -> bool

Returns whether the representation is inline.

pub unsafe fn out_of_line_offset(&self) -> isize

Returns the offset of the representation.

Safety

The internal representation must be out-of-line.

pub fn as_ptr(&self) -> *const u8

Returns a pointer to the bytes of the string.

pub fn as_mut_ptr(&mut self) -> *mut u8

Returns a mutable pointer to the bytes of the string.

pub fn len(&self) -> usize

Returns the length of the string.

pub fn is_empty(&self) -> bool

Returns whether the string is empty.

pub fn as_str_ptr(&self) -> *const str

Returns a pointer to the string as a str.

pub fn bytes(&self) -> &[u8]

Returns a slice of the bytes of the string.

pub fn bytes_mut(&mut self) -> &mut [u8]

Returns a mutable slice of the bytes of the string.

pub fn as_str(&self) -> &str

Returns a reference to the string as a str.

pub fn as_mut_str(&mut self) -> &mut str

Returns a mutable reference to the string as a str.

pub unsafe fn emplace_inline(value: &str, out: *mut Self)

Emplaces a new inline representation for the given str.

Safety

The length of str must be less than or equal to INLINE_CAPACITY.

out must point to a valid location to write the inline representation.

pub unsafe fn emplace_out_of_line( value: &str, pos: usize, target: usize, out: *mut Self, )

Emplaces a new out-of-line representation for the given str.

Safety

The length of str must be greater than INLINE_CAPACITY.

pos must be the location of the representation within the archive.

target must be the location of the serialized bytes of the string.

out must point to a valid location to write the out-of-line representation.