rkyv

Trait Deserialize

pub trait Deserialize<T, D: Fallible + ?Sized> {
    // Required method
    fn deserialize(&self, deserializer: &mut D) -> Result<T, D::Error>;
}

Converts a type back from its archived form.

Some types may require specific deserializer capabilities, such as Rc and Arc. In these cases, the deserializer type D should be bound so that it implements traits that provide those capabilities (e.g. SharedDeserializeRegistry).

This can be derived with Deserialize.

Required Methods

fn deserialize(&self, deserializer: &mut D) -> Result<T, D::Error>

Deserializes using the given deserializer