rkyv::validation::shared

Trait SharedContext

pub trait SharedContext<E = <Self as Fallible>::Error> {
    // Required methods
    fn start_shared(
        &mut self,
        address: usize,
        type_id: TypeId,
    ) -> Result<ValidationState, E>;
    fn finish_shared(
        &mut self,
        address: usize,
        type_id: TypeId,
    ) -> Result<(), E>;
}

A context that can validate shared archive memory.

Shared pointers require this kind of context to validate.

Required Methods

fn start_shared( &mut self, address: usize, type_id: TypeId, ) -> Result<ValidationState, E>

Starts validating the value associated with the given address.

Returns an error if the value associated with the given address was started with a different type ID.

fn finish_shared(&mut self, address: usize, type_id: TypeId) -> Result<(), E>

Finishes validating the value associated with the given address.

Returns an error if the given address was not pending.