Struct rkyv::with::With

pub struct With<F: ?Sized, W> { /* private fields */ }

A transparent wrapper for archived fields.

This is used by the #[with(...)] attribute in the Archive macro to create transparent serialization wrappers. Those wrappers leverage ArchiveWith to change how the type is archived, serialized, and deserialized.

When a field is serialized, a reference to the field (i.e. &T) can be cast to a reference to a wrapping With (i.e. With<T, Wrapper>) and serialized instead. This is safe to do because With is a transparent wrapper and is shaped exactly the same as the underlying field.

Example

use rkyv::{Archive, with::Inline};

#[derive(Archive)]
struct Example<'a> {
    // This will archive as if it were With<&'a i32, Inline>. That will delegate the archival
    // to the ArchiveWith implementation of Inline for &T.
    #[with(Inline)]
    a: &'a i32,
}

Implementations

impl<F: ?Sized, W> With<F, W>

pub fn cast(field: &F) -> &With<F, W>

Casts a With reference from a reference to the underlying field.

This is always safe to do because With is a transparent wrapper.

impl<F, W> With<F, W>

pub fn into_inner(self) -> F

Unwraps a With into the underlying field.