Trait rkyv::with::SerializeWith

pub trait SerializeWith<F: ?Sized, S: Fallible + ?Sized>: ArchiveWith<F> {
    fn serialize_with(
        field: &F, 
        serializer: &mut S
    ) -> Result<Self::Resolver, S::Error>;
}

A variant of Serialize that works with With wrappers.

Required methods

fn serialize_with(
field: &F,
serializer: &mut S
) -> Result<Self::Resolver, S::Error>

Serializes the field type F using the given serializer.